requirements.txt
# Development dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Email
aiosmtplib==3.0.1

# Export/Import
openpyxl==3.1.2
reportlab==4.0.9
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development tools
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0

# Monitoring (optional)
python-json-logger==2.0.7
